+++ src/test/java/id/grocery/tunas/product/ProductRepositoryTest.java
@@ -9,6 +9,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
@@ -33,7 +34,7 @@ class ProductRepositoryTest {
         product.setId(UUID.fromString("0b589615-f910-11eb-936c-41a335bdee2c"));
         product.setName("Broccoli");
         product.setDescription("Green vegetable, good for eyes");
-        product.setPrice(1100L);
+        product.setPrice(BigDecimal.valueOf(1100L));
         product.setWeight(1500);
         product.setPerUnit(100);
         product.setCategory(category);
